common & core & app-ktor: replace CardRepositories & DictionaryRepositories with AppRepositories
