+++ app-ktor/src/main/kotlin/api/controllers/DictionaryController.kt
@@ -1,7 +1,7 @@
 package com.gitlab.sszuev.flashcards.api.controllers
 
+import com.gitlab.sszuev.flashcards.AppRepositories
 import com.gitlab.sszuev.flashcards.DictionaryContext
-import com.gitlab.sszuev.flashcards.DictionaryRepositories
 import com.gitlab.sszuev.flashcards.api.services.DictionaryService
 import com.gitlab.sszuev.flashcards.api.v1.models.BaseRequest
 import com.gitlab.sszuev.flashcards.api.v1.models.CreateDictionaryRequest
@@ -22,7 +22,7 @@ private val logger: ExtLogger = logger("com.gitlab.sszuev.flashcards.api.control
 
 suspend fun ApplicationCall.getAllDictionaries(
     service: DictionaryService,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     contextConfig: ContextConfig
 ) {
     execute<GetAllDictionariesRequest>(
@@ -37,7 +37,7 @@ suspend fun ApplicationCall.getAllDictionaries(
 
 suspend fun ApplicationCall.createDictionary(
     service: DictionaryService,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     contextConfig: ContextConfig
 ) {
     execute<CreateDictionaryRequest>(
@@ -52,7 +52,7 @@ suspend fun ApplicationCall.createDictionary(
 
 suspend fun ApplicationCall.deleteDictionary(
     service: DictionaryService,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     contextConfig: ContextConfig
 ) {
     execute<DeleteDictionaryRequest>(
@@ -67,7 +67,7 @@ suspend fun ApplicationCall.deleteDictionary(
 
 suspend fun ApplicationCall.downloadDictionary(
     service: DictionaryService,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     contextConfig: ContextConfig
 ) {
     execute<DownloadDictionaryRequest>(
@@ -82,7 +82,7 @@ suspend fun ApplicationCall.downloadDictionary(
 
 suspend fun ApplicationCall.uploadDictionary(
     service: DictionaryService,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     contextConfig: ContextConfig
 ) {
     execute<UploadDictionaryRequest>(
@@ -97,7 +97,7 @@ suspend fun ApplicationCall.uploadDictionary(
 
 private suspend inline fun <reified R : BaseRequest> ApplicationCall.execute(
     operation: DictionaryOperation,
-    repositories: DictionaryRepositories,
+    repositories: AppRepositories,
     logger: ExtLogger,
     contextConfig: ContextConfig,
     noinline exec: suspend DictionaryContext.() -> Unit,
